+++ src/pages/PagePhotoOps.jsx
@@ -1,31 +1,69 @@
 import React, { useState, useEffect } from "react";
+import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import Heading from '../components/Heading';
 import Footer from "../components/Footer";
 import setBodyColor from '../setBodyColor'
 import Tile from '../components/Tile';
+import ListPhotoOps from "../components/ListPhotoOps";
 import './PagePhotoOps.scss';
-import data from "../data/photoopdata.json";
+
+//import data from "../data/photoopdata.json";
+import dataClient from "../data/sanityClient";
 
 const PagePhotoOps = () => {
-    const [searchTerm, setSearchTerm] = useState("");
-    const [filteredData, setFilteredData] = useState(data);
+    const [searchTerm, setSearchTerm] = useState("");    
+    const [photoOpData, setPhotoOpData] = useState([]);
+    const [filteredData, setFilteredData] = useState(photoOpData);
 
     useEffect(() => {
-        const filtered = data
-          .map((type) => ({
-            ...type,
-            guests: type.guests.filter((guest) =>
-              Object.values(guest)
-                .map((value) => value.toString().toLowerCase())
-                .some((element) => element.includes(searchTerm.toLowerCase()))
-            ),
-          }))
-          .filter((type) => type.guests.length > 0);
-    
-        setFilteredData(filtered);
+        let query = `*[_type == "pmcPhotoOp" && pmcYear == 2024] | order(type desc, guests[0]->name asc, groupName asc) {
+            _id,
+            pmcYear,
+            type,
+            groupName,
+            rate,
+            guests[]->{name},
+            photoOpTime
+        }`;
+        dataClient
+            .fetch(query)
+            .then((data) => { setPhotoOpData(data); setFilteredData(data); })
+            .catch((e) => { console.log(e); });
 
         document.title = 'Photo Ops - Power Morphicon Planner';
         window.scrollTo(0, 0);
+    },[]);
+
+    useEffect(() => {
+        const lowerSearchTerm = searchTerm.toLowerCase();
+        
+        if(photoOpData) {
+            const filteredPhotoOps = photoOpData.filter((photoOp) => {
+                const photoOpValues = [
+                    photoOp.groupName ? photoOp.groupName.toLowerCase() : '',
+                    ...(photoOp.guests ? photoOp.guests.map((guest) => guest.name.toLowerCase()) : [])
+                ];
+                return photoOpValues.some((value) => value.includes(lowerSearchTerm));
+            });
+            setFilteredData(filteredPhotoOps);
+        } else {
+            setFilteredData(photoOpData);
+        }
+
+        // const filtered = data
+        //   .map((type) => ({
+        //     ...type,
+        //     guests: type.guests.filter((guest) =>
+        //       Object.values(guest)
+        //         .map((value) => value.toString().toLowerCase())
+        //         .some((element) => element.includes(searchTerm.toLowerCase()))
+        //     ),
+        //   }))
+        //   .filter((type) => type.guests.length > 0);
+    
+        //setFilteredData(filtered);
+
+        
       }, [searchTerm]);
 
     const handleInputChange = (e) => {
@@ -64,62 +102,37 @@ const PagePhotoOps = () => {
                 </p>
                 <input className="photo-ops-searcher" id="photoOpSearcher" type="text" placeholder="Search by guest or group name" value={searchTerm} onChange={handleInputChange} />
 
-                { filteredData.length == 0 ? (<p className="no-data">No photo ops matched your search for <b>"{searchTerm}."</b></p>) : "" }
-
-                { filteredData.map((type, index) => (
-                    <Tile key={index} className='results photo-ops section'>
-                        <div className="heading">
-                            <h3>{type.type}</h3>
-                        </div>
-                        <table className="photo-ops-list" cellSpacing="0">
-                            <tbody>
-                                {type.guests.map((guest, idx) => (
-                                    <tr key={idx}>
-                                    <td className="details">
-                                        { guest.group_name ? (
-                                            <>
-                                            <h3>{ guest.group_name }</h3>
-                                            <p className="description">{ guest.name }</p>
-                                            </>
-                                        ) : (
-                                            <h3>{ guest.name }</h3>
-                                        ) }
-                                        {guest.rate ? (<p className="rate"></p>) : ''} 
-                                    </td>
-                                    <td className="time">
-                                        {guest.fri ? guest.fri : (
-                                            <span className="no-data">
-                                                &times;
-                                            </span>
-                                        )}
-                                        <p><span>Fri</span></p>
-                                    </td>
-                                    <td className="time">
-                                        {guest.sat ? guest.sat : (
-                                            <span className="no-data">
-                                                &times;
-                                            </span>
-                                        )}
-                                        <p><span>Sat</span></p>
-                                    </td>
-                                    <td className="time">
-                                        {guest.sun ? guest.sun : (
-                                            <span className="no-data">
-                                                &times;
-                                            </span>
-                                        )}
-                                        <p><span>Sun</span></p>
-                                    </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
-                    </Tile>
-                )) }
+                <Tabs className='photoop-tabs' selectedTabClassName='active'>
+                    <TabList>
+                        <Tab><span>Solo</span></Tab>
+                        <Tab><span>In-Costume</span></Tab>
+                        <Tab><span>Group</span></Tab>
+                    </TabList>
+                    <TabPanel>
+                        { filteredData && filteredData.filter((photoOp) => photoOp.type === 'solo').length > 0 ? (
+                            <ListPhotoOps data={filteredData.filter((photoOp) => photoOp.type === 'solo')} />
+                        ) : (
+                            <div className="no-data">No matching solo photo ops for <b>"{searchTerm}."</b></div>
+                        )}
+                    </TabPanel>
+                    <TabPanel>
+                        { filteredData && filteredData.filter((photoOp) => photoOp.type === 'costume').length > 0 ? (
+                            <ListPhotoOps data={filteredData.filter((photoOp) => photoOp.type === 'costume')} />
+                        ) : (
+                            <div className="no-data">No matching in-costume photo ops for <b>"{searchTerm}."</b></div>
+                        )}
+                    </TabPanel>
+                    <TabPanel>
+                        { filteredData && filteredData.filter((photoOp) => photoOp.type === 'group').length > 0 ? (
+                            <ListPhotoOps data={filteredData.filter((photoOp) => photoOp.type === 'group')} />
+                        ) : (
+                            <div className="no-data">No matching group photo ops for <b>"{searchTerm}."</b></div>
+                        )}
+                    </TabPanel>
+                </Tabs>
                 <Tile className='disclaimer'>
                     <Footer />
                 </Tile>
-
             </div>
         </div>
     </div>