+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-09-15
+### Changed 
+- Changed datasource from JSON to Sanity.io
+- Minor revamp to Panels and Photo Ops layout
+### Added
+- Guest+Bio module
+
 ## [0.3.1] - 2024-09-14
 ### Changed
 - Darkened the Tile colour schemes